Move and rename UnreadWidgetData
